BC-8754 - rename KebabMenuActionDownload to KebabMenuActionDownloadFiles and update related components
